+++ src/main/kotlin/com/beeroadam/repository/AttributeRepository.kt
@@ -1,0 +1,7 @@
+package com.beeroadam.repository
+
+import com.beeroadam.entity.Attribute
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface AttributeRepository: JpaRepository<Attribute, Long> {
+}