+++ src/main/kotlin/com/beeroadam/repository/ProductDetailsRepository.kt
@@ -1,0 +1,7 @@
+package com.beeroadam.repository
+
+import com.beeroadam.entity.ProductDetails
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ProductDetailsRepository: JpaRepository<ProductDetails, Long> {
+}